Added parachute model
Added the chute property in Mjollnir. The parachute model assumes the drag force from the chutes acts in the direction opposite to the rocket's velocity vector. The force is applied to a point near the nosecone. The drogue chute is set to deploy when the rocket starts descending from apogee. When a set altitude is achieved, the main chute is deployed (replacing the drogue).